[Style] seperate Model from function
