+++ HappyAnding/HappyAnding/ViewModel/CheckUpdateVersion.swift
@@ -10,19 +10,6 @@ import SwiftUI
 import Combine
 import FirebaseFirestore
 
-struct Version: Codable {
-    var latestVersion: String
-    var minimumVersion: String
-    var description: String
-    var title: String
-    
-    enum CodingKeys: String, CodingKey {
-        case latestVersion = "latest_version"
-        case minimumVersion = "minimum_version"
-        case title, description
-    }
-}
-
 class CheckUpdateVersion {
     
     static let share = CheckUpdateVersion()
